requirements.txt
openai>=1.14.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
openpyxl>=3.1.2,<4.0.0
pytest>=8.1.0,<9.0.0
